Cleanup mappers by adding a mapper decorator
I am not sure this is a good change. The advantage is that the mapper
decorator consolidates a lot of the error handling. The actual mapper
functions are a lot cleaner and easier to reason about.

The not-so-nice bit is that the decorator adds a couple of
kwargs (error, default) to the mapper. This is confusing since these
kwargs are not in the mapper's signature.

As a side effect, _get_measure was cleaned up.