Fix application of multiple mappers.
apply_mapper can now take a tuple of mappers.
It will apply them sequentially. Mappers are expected to return []
if no mapping is possible. A list of results e.g. [[], ["lon"]] is then
unpacked to ["lon"] and returned. If none of the mappers apply or if
more than one mapper can apply, an error is raised.