Changed coord_criteria approach
These changes are in support of updating the `_get_axis_coord` logic to be a bit more generalized so the same logic can be used to find axis coordinates and to pull out variables (that is a future step).

The coordinate_criteria dictionary in criteria.py is now inverted so the keys are axes and coord names with values of valid attribute types and values.

The function `_get_axis_coord()` was updated to try to account for this change in the `coordinate_criteria` dictionary though it may not be correct quite yet. The tests did not all pass.