Changed logic back so name matches exactly
Changed the logic back in `_get_custom_criteria` so that "name" has to match exactly since sometimes variable names are extensions of each other and it is reasonable to assume they need to match exactly. For example, this way "wind_speed" and "wind_speed_of_gust" can both be identified by name distinctly.